+++ src/cloudflare-workers/index.js
@@ -1,7 +1,7 @@
 /**
  * This file is part of Athene2 Assets.
  *
- * Copyright (c) 2017-2018 Serlo Education e.V.
+ * Copyright (c) 2017-2019 Serlo Education e.V.
  *
  * Licensed under the Apache License, Version 2.0 (the "License")
  * you may not use this file except in compliance with the License
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *
- * @copyright Copyright (c) 2013-2018 Serlo Education e.V.
+ * @copyright Copyright (c) 2013-2019 Serlo Education e.V.
  * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License 2.0
  * @link      https://github.com/serlo-org/athene2-assets for the canonical source repository
  */
@@ -25,14 +25,32 @@ addEventListener('fetch', event => {
 
 export async function handleRequest(request) {
   const response =
-    (await handleSemanticAssetsFilenames(request)) || (await fetch(request))
+    (await handleRedirects(request)) ||
+    (await handleSemanticAssetsFilenames(request)) ||
+    (await fetch(request))
 
   return response
 }
 
+async function handleRedirects(request) {
+  const { url } = request
+
+  if (/^https?:\/\/start.serlo.org/.test(url)) {
+    return Response.redirect(
+      'https://docs.google.com/document/d/1qsgkXWNwC-mcgroyfqrQPkZyYqn7m1aimw2gwtDTmpM/'
+    )
+  }
+}
+
 async function handleSemanticAssetsFilenames(request) {
-  const re = /^https:\/\/assets.serlo.org\/(legacy\/|meta\/|)((?!legacy|meta)\w+)\/([\w\-+]+)\.(\w+)$/
-  const match = request.url.match(re)
+  const { url } = request
+
+  if (/^https:\/\/assets.serlo.org\/meta\//.test(url)) {
+    return null
+  }
+
+  const re = /^https:\/\/assets.serlo.org\/(legacy\/|)((?!legacy)\w+)\/([\w\-+]+)\.(\w+)$/
+  const match = url.match(re)
 
   if (!match) {
     return null