+++ src/main/modules/recaptcha.js
@@ -1,7 +1,7 @@
 /**
  * This file is part of Athene2 Assets.
  *
- * Copyright (c) 2017-2018 Serlo Education e.V.
+ * Copyright (c) 2017-2019 Serlo Education e.V.
  *
  * Licensed under the Apache License, Version 2.0 (the "License")
  * you may not use this file except in compliance with the License
@@ -15,12 +15,34 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *
- * @copyright Copyright (c) 2013-2018 Serlo Education e.V.
+ * @copyright Copyright (c) 2013-2019 Serlo Education e.V.
  * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License 2.0
  * @link      https://github.com/serlo-org/athene2-assets for the canonical source repository
  */
-import createEditorPlugins, { defaultPlugin } from '@serlo-org/editor-plugins'
+import $ from 'jquery'
 
-export default createEditorPlugins
+const bindCallback = (index, $form) => {
+  window[`reCaptchaCallbacks_${index}`] = createSubmitCallback($form)
+  $('button.g-recaptcha', $form).attr(
+    'data-callback',
+    `reCaptchaCallbacks_${index}`
+  )
+}
 
-export { defaultPlugin }
+const createSubmitCallback = $form => () => {
+  $form.submit()
+}
+
+const injectReCaptchaScript = () => {
+  $('body').append(
+    '<script src="https://www.google.com/recaptcha/api.js" async defer></script>'
+  )
+}
+
+$.fn.ReCaptcha = function() {
+  const $forms = $(this)
+  $forms.each(bindCallback)
+  if ($forms.length > 0) {
+    injectReCaptchaScript()
+  }
+}