+++ CHANGELOG.md
@@ -8,6 +8,44 @@ All notable changes to this project will be documented in this file. The format 
 
 - Improved content styles
 
+## [3.6.0] - 2019-01-22
+
+### Added
+
+- Content API: notify parent window of actual content dimensions
+- Editor: shortcuts for undo/redo
+
+### Fixed
+
+- Editor: various fixes
+
+## [3.5.0] - 2019-01-16
+
+### Added
+
+- H5P.com editor plugin
+
+### Fixed
+
+- Upload `yarn.lock` with Google Cloud Function for consistent dependencies
+
+## [3.4.0] - 2019-01-14
+
+### Added
+
+- Lazy load ReCAPTCHA
+- Redirect from https://start.serlo.org to team overview document
+
+### Changed
+
+- Update @serlo/editor to v0.4.0
+
+### Fixed
+
+- Correct heading styles in new editor
+- Set max filesize limit of old image upload to 2MB
+- Handle rewriting of https://assets.serlo.org/meta correctly
+
 ## [3.3.0] - 2018-12-02
 
 ### Changed
@@ -97,7 +135,10 @@ All notable changes to this project will be documented in this file. The format 
 
 ## [2.0.4] - 2018-09-28
 
-[unreleased]: https://github.com/serlo-org/athene2-assets/compare/3.3.0...HEAD
+[unreleased]: https://github.com/serlo-org/athene2-assets/compare/3.6.0...HEAD
+[3.6.0]: https://github.com/serlo-org/athene2-assets/compare/3.5.0...3.6.0
+[3.5.0]: https://github.com/serlo-org/athene2-assets/compare/3.4.0...3.5.0
+[3.4.0]: https://github.com/serlo-org/athene2-assets/compare/3.3.0...3.4.0
 [3.3.0]: https://github.com/serlo-org/athene2-assets/compare/3.2.2...3.3.0
 [3.2.2]: https://github.com/serlo-org/athene2-assets/compare/3.2.1...3.2.2
 [3.2.1]: https://github.com/serlo-org/athene2-assets/compare/3.2.0...3.2.1